+++ go-client-sdk/pkg/models/shared/simpleobjectcamelcase.go
@@ -120,7 +120,7 @@ func (s SimpleObjectCamelCase) MarshalJSON() ([]byte, error) {
 }
 
 func (s *SimpleObjectCamelCase) UnmarshalJSON(data []byte) error {
-	if err := utils.UnmarshalJSON(data, &s, "", false, false); err != nil {
+	if err := utils.UnmarshalJSON(data, &s, "", false, true); err != nil {
 		return err
 	}
 	return nil